added fontend stuff
