+++ frontend/flask/Blueprints/Assesment.py
@@ -8,7 +8,7 @@ from flask import (
 )
 from flask import Blueprint
 from datetime import datetime
-from frontend.flask.Forms import ScholarshipApplicationForm  # Ensure correct import path
+from Forms import ScholarshipApplicationForm  # Ensure correct import path
 
 assesment_bp = Blueprint("assesment", __name__, url_prefix="/assessment")  # Corrected spelling to "assessment"
